Fix #46
