+++ .eslintrc.json
@@ -6,6 +6,7 @@
     "no-param-reassign": ["error", { "props": false }],
     "new-cap": ["error", { "properties": false }],
     "no-underscore-dangle": "off",
-    "max-len": "off"
+    "max-len": "off",
+    "no-console": "off"
   }
 }
