+++ lib/todo/fetch-data.js
@@ -12,7 +12,7 @@ const gitlab = axios.create({
 	},
 });
 
-const github = axios.create({
+const github = () => axios.create({
 	baseURL: 'https://api.github.com/',
 	timeout: 1500,
 	headers: {
@@ -37,10 +37,12 @@ const fetchGitlab = new Promise((resolve, reject) => {
 	}).catch(reject);
 });
 
-const fetchGithub = new Promise((resolve, reject) => {
-	github.get('/issues')
-	.then((data) => resolve(data.data)).catch(reject);
-});
+const fetchGithub = () => github().get('/issues').then(data => data.data);
+
+// const fetchGithub = new Promise((resolve, reject) => {
+// 	github.get('/issues')
+// 	.then((data) => resolve(data.data)).catch(reject);
+// });
 
 const fetchTrello = new Promise((resolve, reject) => {
 	trello.get('/1/members/me/cards', (err, cards) => {
