+++ package.json
@@ -37,6 +37,7 @@
     "color-namer": "^1.1.0",
     "commander": "^2.9.0",
     "ini": "^1.3.4",
+    "inquirer": "^3.0.6",
     "lodash": "^4.17.4",
     "moment": "^2.18.0",
     "node-trello": "^1.1.2",
